luping0702

+++ app/controllers/sv_cards_controller.rb
@@ -19,27 +19,31 @@ class SvCardsController < ApplicationController   #优惠卡控制器
     card_name = params[:card_name].strip
     card_description = params[:card_description]
     img = params[:card_url]      #获取上传的图片
-    if card_type == 1                                       #如果是储值卡
-      started_money = params[:started_money].to_f
-      ended_money = params[:ended_money].to_f
-      sv_card.update_attributes(:name => card_name, :types => card_type, :price => started_money, :description => card_description)
-      if sv_card.save
-        SvcardProdRelation.create(:sv_card_id => sv_card.id, :base_price => started_money, :more_price => ended_money)        
+    if SvCard.where(["types= ? and name = ?", card_type, card_name]).blank?
+      if card_type == 1                                       #如果是储值卡
+        started_money = params[:started_money].to_f
+        ended_money = params[:ended_money].to_f
+        sv_card.update_attributes(:name => card_name, :types => card_type, :price => started_money, :description => card_description)
+        if sv_card.save
+          SvcardProdRelation.create(:sv_card_id => sv_card.id, :base_price => started_money, :more_price => ended_money)
+        end
+      elsif card_type == 0                                        #如果是打折卡
+        discount = params[:discount_value]
+        price = params[:discount_price].to_f
+        sv_card.update_attributes(:name => card_name, :types => card_type,:discount => discount, :price => price, :description => card_description)
+        sv_card.save
       end
-    elsif card_type == 0                                        #如果是打折卡
-      discount = params[:discount_value]
-      price = params[:discount_price].to_f
-      sv_card.update_attributes(:name => card_name, :types => card_type,:discount => discount, :price => price, :description => card_description)
-      sv_card.save
+      begin
+        url = SvCard.upload_img(img, sv_card.id, Constant::SVCARD_PICS, Constant::STORE_ID, Constant::SVCARD_PICSIZE)
+        sv_card.update_attribute("img_url", url)
+        flash[:notice] = "创建成功!"
+      rescue
+        flash[:notice] = "图片上传失败!"
+      end
+    else
+      flash[:notice] = "创建失败，已有同名的优惠卡!"
     end
-     begin
-          url = SvCard.upload_img(img, sv_card.id, Constant::SVCARD_PICS, Constant::STORE_ID, Constant::SVCARD_PICSIZE)
-          sv_card.update_attribute("img_url", url)
-          flash[:notice] = "创建成功!"         
-        rescue
-          flash[:notice] = "图片上传失败!"
-      end
-       redirect_to sv_cards_path
+    redirect_to sv_cards_path
   end
 
   def update    #更新优惠卡
@@ -48,36 +52,40 @@ class SvCardsController < ApplicationController   #优惠卡控制器
     type = sc.types
     description = params[:edit_card_description]
     img = params[:edit_card_url]
-    if type == 1
-      started_money = params[:edit_started_money].to_f
-      ended_money = params[:edit_ended_money].to_f
-      if sc.update_attributes(:name => name, :price => started_money, :description => description)
-        SvcardProdRelation.destroy_all("sv_card_id = #{sc.id}")
-        SvcardProdRelation.create(:sv_card_id => sc.id, :base_price => started_money, :more_price => ended_money)
-        if !img.nil?
-          begin
-            new_url = SvCard.upload_img(img, sc.id, Constant::SVCARD_PICS, Constant::STORE_ID, Constant::SVCARD_PICSIZE)
-            sc.update_attribute("img_url", new_url)
-          rescue
-            flash[:notice] ="图片更新失败！"
+    if SvCard.where(["id != ? and types = ? and name = ?", sc.id, type, name]).blank?
+      if type == 1
+        started_money = params[:edit_started_money].to_f
+        ended_money = params[:edit_ended_money].to_f
+        if sc.update_attributes(:name => name, :price => started_money, :description => description)
+          SvcardProdRelation.destroy_all("sv_card_id = #{sc.id}")
+          SvcardProdRelation.create(:sv_card_id => sc.id, :base_price => started_money, :more_price => ended_money)
+          if !img.nil?
+            begin
+              new_url = SvCard.upload_img(img, sc.id, Constant::SVCARD_PICS, Constant::STORE_ID, Constant::SVCARD_PICSIZE)
+              sc.update_attribute("img_url", new_url)
+            rescue
+              flash[:notice] ="图片更新失败！"
+            end
           end
+          flash[:notice] = "更新成功!"
         end
-        flash[:notice] = "更新成功!"
-      end
-    elsif type == 0
-      discount = params[:edit_discount_value]
-      price = params[:edit_discount_price]
-      if sc.update_attributes(:name => name,:description => description, :discount => discount, :price => price)
-        if !img.nil?
-          begin
-            new_url = SvCard.upload_img(img, sc.id, Constant::SVCARD_PICS, Constant::STORE_ID, Constant::SVCARD_PICSIZE)
-            sc.update_attribute("img_url", new_url)      
-          rescue
-            flash[:notice] ="图片更新失败！"
+      elsif type == 0
+        discount = params[:edit_discount_value]
+        price = params[:edit_discount_price]
+        if sc.update_attributes(:name => name,:description => description, :discount => discount, :price => price)
+          if !img.nil?
+            begin
+              new_url = SvCard.upload_img(img, sc.id, Constant::SVCARD_PICS, Constant::STORE_ID, Constant::SVCARD_PICSIZE)
+              sc.update_attribute("img_url", new_url)
+            rescue
+              flash[:notice] ="图片更新失败！"
+            end
           end
+          flash[:notice] = "更新成功!"
         end
-        flash[:notice] = "更新成功!"
       end
+    else 
+      flash[:notice] = "更新失败,已有同名的优惠卡!"
     end
     redirect_to request.referer
   end
